+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+## [0.3.0] - 2020-05-25
+
 ### Added
 
 - Support Aave deposit and redeem.